hash function improved + moved some stuff to Sharptoken.lib
