+++ SharpToken/Lib/BytePairEncodingCore.cs
@@ -1,12 +1,11 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 
-namespace SharpToken
+namespace SharpToken.Lib
 {
     public class BytePairEncodingCore
     {
@@ -225,24 +224,6 @@ namespace SharpToken
                 var key = inputBytes.Skip(pair.Start).Take(pair.End - pair.Start).ToArray();
                 return bytePairRanks[key];
             });
-        }
-    }
-
-    internal sealed class ByteArrayEqualityComparer : IEqualityComparer<byte[]>
-    {
-        public bool Equals(byte[] x, byte[] y)
-        {
-            if (x == null || y == null)
-            {
-                return false;
-            }
-
-            return ReferenceEquals(x, y) || StructuralComparisons.StructuralEqualityComparer.Equals(x, y);
-        }
-
-        public int GetHashCode(byte[] obj)
-        {
-            return StructuralComparisons.StructuralEqualityComparer.GetHashCode(obj);
         }
     }
 }
