+++ SharpToken/Lib/Encoding.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
+using SharpToken.Lib;
 
 namespace SharpToken
 {
